requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
openai>=1.12.0
anthropic>=0.18.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
plotly>=5.17.0
pandas>=2.0.0

